nvkms: Implement CRTC_X/Y for primary planes
This is used by Gamescope for centring the game's buffer.

Adds a check to see if the `supportsWindowMode` cap is supported
for the main plane (EVO3+) and implements CRTC_X/Y for
NVKMS_MAIN_LAYER if it is supported.

This was not checked before, so CRTC_X could be specified on the main
layer, and the atomic commit would succeed, resulting in incorrect
output.