nvidia: Remove unsupported GPU check
This check solely gets in the way of developer and end-users testing
and provides no benefit to either.

The check as it stands only prints a selection of unhelpful and
unrelated errors to TTY, which appears like a broken driver install.
(which is not a good look!)

Multiple distros are already shipping with this useless check
defaulted to off, so keeping it around is just dead weight
that's getting in people's way and an annoying 'gotcha'.

It should really just be removed.