Fix NV2080_CTRL_CMD_GPU_GET_PID_INFO don't work correctly in container
When call nvmlDeviceGetComputeRunningProcesses_v3() in container,
the pid returned is convert by _gpuConvertPid, and will get the
pid which is inside the namespace, but not
the globle one. When user use the pid to call
NV2080_CTRL_CMD_GPU_GET_PID_INFO, we do not do that fasion
translation, just compare it with ProcID, this leads users
get NULL data inside container.

To fix this issue, just keep the same, do the translation,
so that apps could get the pid data.

Signed-off-by: Michael Qiu <qdy220091330@gmail.com>